decouple variant projection handler
