+++ be/src/formats/file_writer.h
@@ -18,6 +18,7 @@
 #include <functional>
 #include <future>
 #include <map>
+#include <optional>
 #include <string>
 #include <vector>
