+++ be/src/formats/parquet/group_reader.h
@@ -14,8 +14,6 @@
 
 #pragma once
 
-#include <cctz/time_zone.h>
-
 #include <atomic>
 #include <memory>
 #include <optional>
@@ -27,7 +25,6 @@
 
 #include "cache/scan/shared_buffered_input_stream.h"
 #include "column/column_access_path.h"
-#include "column/variant_path_parser.h"
 #include "column/vectorized_fwd.h"
 #include "common/global_types.h"
 #include "common/object_pool.h"
@@ -52,6 +49,7 @@ class THdfsScanRange;
 
 namespace parquet {
 class FileMetaData;
+class VariantProjectionHandler;
 } // namespace parquet
 struct TypeDescriptor;
 } // namespace starrocks
@@ -128,6 +126,8 @@ struct GroupReaderParam {
 };
 
 class GroupReader {
+    friend class VariantProjectionHandler;
+
 public:
     GroupReader(GroupReaderParam& param, int row_group_number, SkipRowsContextPtr skip_rows_ctx,
                 int64_t row_group_first_row);
@@ -135,8 +135,6 @@ public:
                 int64_t row_group_first_row, int64_t row_group_first_row_id);
     ~GroupReader();
 
-    // init used to init column reader, and devide active/lazy
-    // then we can use inited column collect io range.
     Status init();
     Status prepare();
     const tparquet::ColumnChunk* get_chunk_metadata(SlotId slot_id);
@@ -154,103 +152,54 @@ public:
     bool& get_is_group_filtered() { return _is_group_filtered; }
 
 private:
-    // Describes a virtual sub-path projection from a variant column, e.g. "data.id" in
-    // "SELECT data.id FROM t" where data is a VARIANT column.  The virtual column has no
-    // physical counterpart in the parquet file; it is computed by extracting parsed_path
-    // from the source variant column identified by source_slot_id.
-    struct VariantVirtualProjection {
-        VariantPath parsed_path;    // sub-path to extract, e.g. ["id"]
-        TypeDescriptor target_type; // expected output type of the projected sub-field
-        SlotId source_slot_id;      // slot that holds the source VARIANT column data
-                                    // (either a physical slot or a hidden slot with a negative id)
-    };
-
-    // A VARIANT column that must be read to serve virtual projections but is not itself
-    // present in the output chunk (i.e. not in the user's SELECT list).
-    // A temporary column is allocated in _read_chunk under a synthetic negative slot_id so
-    // that _variant_virtual_projections can reference it without colliding with real slots.
-    // One HiddenVariantSource is created per unique source column name; multiple virtual
-    // columns referencing the same VARIANT source share the same hidden entry.
-    //
-    // is_active: true  → backs at least one virtual column that has a conjunct; must be read
-    //                    before _apply_deferred_variant_conjuncts (included in active_chunk).
-    // is_active: false → backs only projection-only virtual columns; can be read after
-    //                    conjunct evaluation, avoiding reads for rows that are filtered out.
-    struct HiddenVariantSource {
-        SlotId slot_id;                       // synthetic negative id (-1, -2, ...) that identifies this source
-        std::unique_ptr<ColumnReader> reader; // reader for the underlying VARIANT column
-        bool is_active = false;               // active (pre-conjunct) vs lazy (post-conjunct)
-        // true when ALL virtual slots backed by this source are promoted to VariantTypedValueProxy
-        // readers in Phase 2; skip IO and Phase 3 reads for this source entirely.
-        bool fully_promoted = false;
-    };
-
+    // ── Initialization ───────────────────────────────────────────────────────
     void _set_end_offset(int64_t value) { _end_offset = value; }
-
-    // deal_with_pageindex need collect pageindex io range first, it will collect all row groups' io together,
-    // so it should be done in file reader. when reading the current row group, we need first deal_with_pageindex,
-    // and then we can collect io range based on pageindex.
-    Status _deal_with_pageindex();
-
-    void _use_as_dict_filter_column(int col_idx, SlotId slot_id, std::vector<std::string>& sub_field_path);
-    Status _rewrite_conjunct_ctxs_to_predicates(bool* is_group_filtered);
-
-    StatusOr<bool> _filter_chunk_with_dict_filter(ChunkPtr* chunk, Filter* filter);
-    // Evaluates deferred variant virtual conjuncts against active_chunk.
-    // Returns the filter to be applied by the caller; active_chunk is NOT filtered
-    // by this method — the caller merges it with chunk_filter and applies the combined
-    // result once after Phase 4 (in get_next).
-    // An empty filter means there were no deferred conjuncts (all rows pass).
-    // A non-empty all-zero filter means every row was filtered out.
-    // Evaluates deferred variant virtual conjuncts and optionally returns the projected
-    // virtual columns used by those conjuncts.
-    //
-    // projected_chunk:
-    //   - Output chunk keyed by virtual slot id, containing only slots that have deferred
-    //     conjuncts and were projected in this phase.
-    //   - Row order/row count matches `active_chunk` BEFORE Phase-4 combined filter is applied.
-    //     Caller must apply the exact same filter to projected_chunk before passing it to
-    //     _fill_dst_chunk.
-    StatusOr<Filter> _apply_deferred_variant_conjuncts(ChunkPtr& active_chunk, size_t raw_count,
-                                                       ChunkPtr* projected_chunk);
-    Status _align_deferred_projected_chunk_after_filter(const ChunkPtr& active_chunk,
-                                                        const ChunkPtr& deferred_projected_chunk,
-                                                        const Filter& chunk_filter, size_t pre_filter_rows);
-    // Fills output chunk and computes virtual projections. For slots present in
-    // `projected_chunk`, reuse precomputed columns (from deferred conjunct evaluation) to
-    // avoid re-projecting and re-seeking variant paths.
-    Status _fill_dst_chunk(ChunkPtr& active_chunk, const ChunkPtr& projected_chunk, ChunkPtr* chunk);
-    const cctz::time_zone& _get_variant_projection_timezone();
-
     Status _create_column_readers();
     StatusOr<ColumnReaderPtr> _create_reserved_iceberg_column_reader(const SlotDescriptor* slot, int32_t field_id);
     StatusOr<Datum> _get_extended_bigint_value(SlotId slot_id) const;
     StatusOr<ColumnReaderPtr> _create_column_reader(const GroupReaderParam::Column& column);
-    VariantShreddedReadHints _get_variant_shredded_hints(std::string_view column_name) const;
-    Status _prepare_column_readers() const;
-    // Promotes variant virtual columns to VariantTypedValueProxy readers in Phase 2 when the
-    // backing hidden source has _skip_base_payload=true (all paths are scalar typed-value leaves).
-    // Promoted slots are placed in _column_readers instead of _variant_virtual_projections and
-    // the hidden source is marked fully_promoted=true to skip Phase 3 IO/reads.
-    Status _promote_variant_virtual_columns();
-    // Creates a lightweight VIEW chunk of _read_chunk containing the given slot_ids.
-    // Each entry in slot_ids must already exist as a column in _read_chunk. _init_read_chunk()
-    // pre-allocates physical slots and active hidden variant source slots; lazy hidden sources
-    // are allocated separately during Phase 6 of the variant read pipeline.
-    // When include_reserved_fields is true, reserved_field_slots are appended after slot_ids.
-    ChunkPtr _create_read_chunk(const std::vector<SlotId>& slot_ids, bool include_reserved_fields = true);
-    // Extract dict filter columns and conjuncts
     void _process_columns_and_conjunct_ctxs();
-
     bool _try_to_use_dict_filter(const GroupReaderParam::Column& column, ExprContext* ctx,
                                  std::vector<std::string>& sub_field_path, bool is_decode_needed);
-
+    Status _prepare_column_readers() const;
+    Status _deal_with_pageindex();
     Status _init_read_chunk();
+    ChunkPtr _create_read_chunk(const std::vector<SlotId>& slot_ids, bool include_reserved_fields = true);
 
+    // ── Dict-filter support ──────────────────────────────────────────────────
+    void _use_as_dict_filter_column(int col_idx, SlotId slot_id, std::vector<std::string>& sub_field_path);
+    Status _rewrite_conjunct_ctxs_to_predicates(bool* is_group_filtered);
+    StatusOr<bool> _filter_chunk_with_dict_filter(ChunkPtr* chunk, Filter* filter);
+
+    // ── Range read helpers ───────────────────────────────────────────────────
     Status _read_range(const std::vector<int>& read_columns, const Range<uint64_t>& range, const Filter* filter,
                        ChunkPtr* chunk, bool ignore_reserved_field = false);
-
     StatusOr<size_t> _read_range_round_by_round(const Range<uint64_t>& range, Filter* filter, ChunkPtr* chunk);
+
+    // ── get_next() pipeline phases ───────────────────────────────────────────
+    //
+    // 1. Prune deleted rows: applies deletion bitmap to produce chunk_filter.
+    //    Returns true if rows survive; false to skip this range entirely.
+    StatusOr<bool> _prune_deleted_rows(const Range<uint64_t>& r, Filter& chunk_filter, bool& has_filter, size_t count);
+
+    // 2. Read & filter active columns: reads active physical columns and
+    //    evaluates dict / expression filters.  Populates chunk_filter and
+    //    fills active_chunk.  Returns true if rows survive.
+    StatusOr<bool> _read_and_filter_active_columns(const Range<uint64_t>& r, Filter& chunk_filter,
+                                                   ChunkPtr& active_chunk, bool& has_filter, size_t count);
+
+    // 5. Backfill lazy physical columns (no predicates).
+    Status _read_lazy_columns(const Range<uint64_t>& full_range, const Range<uint64_t>& post_filter_range,
+                              const Filter& post_filter, bool has_filter, ChunkPtr& active_chunk);
+
+    // 7. Emit physical columns from active_chunk into the output chunk.
+    Status _emit_physical_columns(ChunkPtr& active_chunk, ChunkPtr* dst);
+
+    // Rebuild ColumnReadOrderCtx from current _active_column_indices.
+    // Called after variant promotion modifies the active column set.
+    void _rebuild_column_read_order_ctx();
+
+    // ── Member variables ─────────────────────────────────────────────────────
 
     // row group meta
     const tparquet::RowGroup* _row_group_metadata = nullptr;
@@ -261,55 +210,12 @@ private:
     // column readers for column chunk in row group
     std::unordered_map<SlotId, std::unique_ptr<ColumnReader>> _column_readers;
 
-    // Maps each virtual-column slot id to its projection descriptor.
-    // Virtual columns (e.g. "data.id") are skipped in the normal fill path and handled
-    // separately in _fill_dst_chunk by extracting the sub-path from the source variant column.
-    std::unordered_map<SlotId, VariantVirtualProjection> _variant_virtual_projections;
-
-    // Maps source variant column name to its hidden reader and temporary slot.
-    // Only populated when the source VARIANT column is not itself in the output (no physical slot).
-    // If the source column IS selected by the user, its physical slot is used directly and no
-    // hidden entry is created, avoiding a double read.
-    std::unordered_map<std::string, HiddenVariantSource> _hidden_variant_sources;
-
-    // Fast reverse-lookup: slot_id → HiddenVariantSource* (pointer into _hidden_variant_sources).
-    // Built in _create_column_readers() after _hidden_variant_sources is fully populated.
-    std::unordered_map<SlotId, HiddenVariantSource*> _hidden_slot_index;
-
-    // Counter for assigning synthetic slot ids to hidden variant sources.
-    // Starts at -1 and decrements so as not to collide with real (non-negative) slot ids.
-    SlotId _next_hidden_slot_id = -1;
-
-    // Slot ids of hidden sources classified as active (needed before conjunct eval) and
-    // lazy (projection-only, deferred until after conjunct eval).
-    // Populated in _process_columns_and_conjunct_ctxs().
-    std::vector<SlotId> _active_hidden_slot_ids;
-    std::vector<SlotId> _lazy_hidden_slot_ids;
-
-    // Unified slot id lists for _create_read_chunk.  Populated by
-    // _process_columns_and_conjunct_ctxs() after the column/conjunct classification.
-    //   _active_slot_ids = physical active slots  +  active hidden variant sources
-    //   _lazy_slot_ids   = physical lazy slots   (hidden lazy sources are read into a
-    //                      temporary chunk in Phase 6, not via _create_read_chunk)
-    std::vector<SlotId> _active_slot_ids;
-    std::vector<SlotId> _lazy_slot_ids;
-
-    // conjunct ctxs that eval after chunk is dict decoded
-    std::vector<ExprContext*> _left_conjunct_ctxs;
-    // variant virtual-column conjuncts can only be evaluated after post-read projection.
-    std::vector<ExprContext*> _deferred_variant_virtual_conjunct_ctxs;
-    // slot ids of virtual columns that have at least one deferred conjunct.
-    // used in _apply_deferred_variant_conjuncts to detect invariant violations.
-    std::unordered_set<SlotId> _deferred_conjunct_slot_ids;
-
-    // Slot ids of virtual variant columns promoted to VariantTypedValueProxy readers in Phase 2.
-    // These slots appear in _column_readers (not _variant_virtual_projections) and participate
-    // in dict filter like normal physical columns.
-    std::unordered_set<SlotId> _promoted_virtual_slots;
+    // ── Variant handler (always present; empty() when no variant columns) ───
+    std::unique_ptr<VariantProjectionHandler> _variant;
 
     // active columns that hold read_col index
     std::vector<int> _active_column_indices;
-    // lazy conlumns that hold read_col index
+    // lazy columns that hold read_col index
     std::vector<int> _lazy_column_indices;
     // load lazy column or not
     bool _lazy_column_needed = false;
@@ -317,16 +223,8 @@ private:
     // dict value is empty after conjunct eval, file group can be skipped
     bool _is_group_filtered = false;
 
-    // Column backing store for each get_next() call.  Initialized once in _init_read_chunk()
-    // and reset at the start of every range iteration.  Holds three categories of columns:
-    //   • Physical slots – one column per _param.read_cols entry plus reserved_field_slots.
-    //     The same ColumnPtr objects are SHARED with active_chunk / lazy_chunk (created via
-    //     _create_read_chunk), so filtering those view-chunks also modifies _read_chunk.
-    //   • Active hidden variant sources – included in active_chunk via _active_slot_ids.
-    //     Filtered automatically by active_chunk->filter() in Phase 4.
-    //   • Lazy hidden variant sources – read into a temporary chunk in Phase 6 (after
-    //     variant conjunct eval) and merged into active_chunk.  _fill_dst_chunk then
-    //     consumes them from active_chunk; they are not accessed via _read_chunk directly.
+    // Column backing store for each get_next() call.  Initialized once in
+    // _init_read_chunk() and reset at the start of every range iteration.
     ChunkPtr _read_chunk;
 
     // param for read row group
@@ -338,15 +236,15 @@ private:
 
     int64_t _end_offset = 0;
 
-    // set to true by _create_column_reader when the FE global-dict map produced a
-    // dict-aware wrapper for at least one slot in this row group; consumed by
-    // _create_column_readers to bump the per-row-group applied counter.
     bool _global_dict_applied_in_group = false;
 
     // columns(index) use as dict filter column
     std::vector<int> _dict_column_indices;
     std::unordered_map<int, std::vector<std::vector<std::string>>> _dict_column_sub_field_paths;
     std::unordered_map<SlotId, std::vector<ExprContext*>> _left_no_dict_filter_conjuncts_by_slot;
+
+    // conjunct ctxs that eval after chunk is dict decoded
+    std::vector<ExprContext*> _left_conjunct_ctxs;
 
     SparseRange<uint64_t> _range;
     SparseRangeIterator<uint64_t> _range_iter;
@@ -357,10 +255,10 @@ private:
     // a flag to reflect prepare() is called
     bool _has_prepared = false;
 
-    // Parsed lazily for VARIANT virtual projections only. Most Parquet readers do not need
-    // this value, so avoid failing unrelated scans on malformed session timezone strings.
-    cctz::time_zone _timezone_obj = cctz::utc_time_zone();
-    bool _timezone_resolved = false;
+    // Unified slot id lists for _create_read_chunk.  Populated by
+    // _process_columns_and_conjunct_ctxs().
+    std::vector<SlotId> _active_slot_ids;
+    std::vector<SlotId> _lazy_slot_ids;
 };
 
 using GroupReaderPtr = std::shared_ptr<GroupReader>;
